+++ Changelog.md
@@ -24,6 +24,7 @@
 - Ensure that browsers cache the correct state of overall comments when marking (#5173)
 - Ensure that graders are shown the correct annotation categories (#5181)
 - Show informative error message if an uploaded criteria yaml file did not contain a "type" key (#5184)
+- Enable content security policies (#5186)
 - Allow for multiple custom validation messages (#5194)
 - Add ability to hold shift to select a range of values in checkbox tables (#5182)
 - Update ssh authorization to be more flexible, secure, and permit a single user to use the same public key for multiple instances (#5199)
@@ -31,7 +32,7 @@
 - Fix bug in Assignment#get_num_marked that caused it to double-count remark and original results (#5205)
 - Fix bug where graders can't see the tests that they run (#5210)
 - Fix bug where graders can't release results on the results page (#5210)
-- Enable content security policies (#5186)
+- Use DOMpurify library to sanitize rendered markdown content (#5211)
 
 ## [v1.11.5]
 - Account for percentage deductions when calculating total marks after deleting a criterion (#5176)